demo/python/requirements.txt
apache-bookkeeper-client==4.17.1
certifi==2024.6.2
click==8.1.7
fastavro==1.9.4
grpcio==1.64.1
joblib==1.4.2
nltk==3.8.1
prometheus_client==0.20.0
protobuf==3.20.3
pulsar-client==3.5.0
pymmh3==0.0.5
pytz==2024.1
ratelimit==2.2.1
regex==2024.5.15
six==1.16.0
tqdm==4.66.4
